chore: organize the distributives
