+++ .gitignore
@@ -3,13 +3,7 @@
 .DS_Store
 node_modules
 
-coverage
-lib
-lib-esm
-lib-fesm
-umd
-typings
-docs
+dist
 
 ## this is generated by `npm pack`
 *.tgz
