+++ .husky/pre-push
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-npm run ts:style && npm run test:only-changed
+npm run eslint && npm run test:only-changed
