+++ e2e/umd-import-min.spec.js
@@ -1,10 +1,9 @@
-const {cold, hot, time} = require('../umd/jest-marbles.min');
+const { cold, hot, time } = require('../dist/umd/jest-marbles.min');
 
 /**
  * Created by Evgeny Barabanov on 05/03/2018.
  */
 describe('Imports test', () => {
-
   it('All the function should exist', () => {
     const c = cold('a|');
     const h = hot('a|');
@@ -15,8 +14,7 @@ describe('Imports test', () => {
   });
 
   it('Should work with value objects', () => {
-    const c = cold('--a-|', {a: {prop: "blah"}});
-    expect(c).toBeObservable(cold('--a-|', {a: {prop: "blah"}}));
+    const c = cold('--a-|', { a: { prop: 'blah' } });
+    expect(c).toBeObservable(cold('--a-|', { a: { prop: 'blah' } }));
   });
-
 });
