+++ e2e/umd-import.spec.js
@@ -1,22 +1,20 @@
-const {cold, hot, time} = require('../umd/jest-marbles');
+const { cold, hot, time } = require('../dist/umd/jest-marbles');
 
 /**
  * Created by Evgeny Barabanov on 05/03/2018.
  */
 describe('Imports test', () => {
+  it('All the function should exist', () => {
+    const c = cold('a|');
+    const h = hot('a|');
+    const t = time('a|');
+    expect(c).not.toBeNull();
+    expect(h).not.toBeNull();
+    expect(t).not.toBeNull();
+  });
 
-    it('All the function should exist', () => {
-        const c = cold('a|');
-        const h = hot('a|');
-        const t = time('a|');
-        expect(c).not.toBeNull();
-        expect(h).not.toBeNull();
-        expect(t).not.toBeNull();
-    });
-
-    it('Should work with value objects', () => {
-      const c = cold('--a-|', {a: {prop: "blah"}});
-      expect(c).toBeObservable(cold('--a-|', {a: {prop: "blah"}}));
-    });
-
+  it('Should work with value objects', () => {
+    const c = cold('--a-|', { a: { prop: 'blah' } });
+    expect(c).toBeObservable(cold('--a-|', { a: { prop: 'blah' } }));
+  });
 });
