+++ package.json
@@ -2,10 +2,10 @@
   "name": "jest-marbles",
   "version": "3.0.6",
   "description": "Marble testing helpers library for RxJs and Jest",
-  "main": "umd/jest-marbles.js",
-  "module": "lib/index.js",
-  "es2015": "lib-esm/index.js",
-  "typings": "typings/index.d.ts",
+  "main": "dist/umd/jest-marbles.js",
+  "module": "dist/lib/index.js",
+  "es2015": "dist/lib-esm/index.js",
+  "typings": "dist/typings/index.d.ts",
   "repository": {
     "type": "git",
     "url": "https://github.com/meltedspark/jest-marbles"
@@ -29,14 +29,14 @@
     "cleanup": "shx rm -rf umd lib lib-esm lib-fesm typings coverage docs",
     "prebuild": "npm run cleanup && npm run verify",
     "postbuild": "npm run e2e",
-    "build": "tsc && tsc --target es2017 --outDir lib-esm && webpack && webpack --env prod",
+    "build": "tsc && tsc --target es2017 --outDir dist/lib-esm && webpack && webpack --env prod",
     "build:fesm:min": "uglifyjs lib-fesm/index.js --compress --mangle --source-map --output lib-fesm/index.min.js",
     "docs": "typedoc -p . --theme minimal --target 'es6' --excludeNotExported --excludePrivate --ignoreCompilerErrors --exclude \"**/src/**/__tests__/*.*\" --out docs src/",
     "test": "jest",
     "test:watch": "npm test -- --watch",
     "test:coverage": "npm test -- --coverage",
     "test:only-changed": "npm test -- --bail --onlyChanged",
-    "eslint": "eslint --ext .ts,.tsx ./",
+    "eslint": "eslint --ext .ts,.tsx ./src ./spec",
     "eslint:fix": "npm run eslint -- --fix",
     "verify": "npm run eslint && npm test",
     "cz": "git-cz",
