+++ tsconfig.json
@@ -16,9 +16,9 @@
     "forceConsistentCasingInFileNames": true,
     "pretty": true,
     "sourceMap": true,
-    "outDir": "lib",
+    "outDir": "dist/lib",
     "declaration": true,
-    "declarationDir": "typings",
+    "declarationDir": "dist/typings",
     "stripInternal": true,
     "skipLibCheck": true,
   },
